==> 7-biomarker/13_r2weight.R <==

# # install/load devtools package
# if(!("devtools" %in% row.names(installed.packages()))){
#     install.packages("devtools")
# }
library("devtools")

# install/load r2weight from GitHub
if(!("r2weight" %in% row.names(installed.packages()))){
    install_github("benkeser/r2weight")
}
library(r2weight)
library(quadprog)
library(tidyverse)

#---------------------------------------------------------------------------
# NOTE: check if function is the right way to get R2
#---------------------------------------------------------------------------
#hmo_fit$out1$univariateR2 gets all univariate R2, so may not need to refit



# Make a function that takes in Y and X, calculates r2weight, comapres the full fit to the models without each predictor one at a time to get importance for combined outcome, and saves the output as a dataframe. 
r2weightFunc <- function(X, Y) {
  
  # call optWeight using simple Super Learner library
  out1 <- optWeight(Y = Y, X = X, SL.library = c("SL.glm","SL.glmnet","SL.randomForest"))
  
  # cross-validated R-squared
  # set verbose = TRUE to see a progress bar
  r2.out1 <- r2_optWeight(out1, Y = Y, X = X)
  
  # Make a dataframe
  results <- data.frame(matrix(ncol = 5, nrow = 7))
  names(results) = c("estFullModel", "estDiff", "CI.l", "CI.h", "p.value")
  
  results $ estFullModel = r2.out1 $ r2
  
  for (i in 1:ncol(X)) {
  # Measure importance of every feature
  out2 <- optWeight(Y = Y, X = X[, -i], SL.library = c("SL.glm","SL.glmnet","SL.randomForest"))
  
  # Get R-squared for combined outcome excluding each feature one at a time
  r2.out2 <- r2_optWeight(out2, Y = Y, X = X[, -i])
  
  # Compare to full fit to get importance for combined outcome
  outDiff.r2 <- r2_diff(r2.out1, r2.out2)
  
  # Save the difference as a data frame
  results[i, 2] = outDiff.r2 $ diff[, 1]
  results[i, 3] = outDiff.r2 $ diff[, 2]
  results[i, 4] = outDiff.r2 $ diff[, 3]
  results[i, 5] = outDiff.r2 $ diff[, 4]
  }
  
  return(list(results=results, out1=out1))
}


# Load datasets and prepare
# HMO Elicit
try(hmoE <- readRDS("/data/imic/data/raw_lab_data/elicit/merged_elicit/hmo.RDS"))
try(hmoE <- readRDS("C:/Users/andre/Downloads/hmo.RDS"))



# Prepare data for analysis
hmoE <- data.frame(hmoE %>%
                  select(c("nlchild_base", "nperson_base", "nrooms_base", 
                            "meducyrs_base", "h2osrcp_base", "cookplac_base",
                            "inctot_base", "DFLac_nmol.mL",
                           "X2.FL_nmol.mL", "X3FL_nmol.mL", "X3.SL_nmol.mL",
                           "X6.SL_nmol.mL", "LNT_nmol.mL", "LNnT_nmol.mL",
                           "LNFP.I_nmol.mL", "LNFP.II_nmol.mL", "LNFP.III_nmol.mL",
                           "LSTb_nmol.mL", "LSTc_nmol.mL", "DFLNT_nmol.mL",
                           "LNH_nmol.mL", "DSLNT_nmol.mL", "FLNH_nmol.mL",
                           "DFLNH_nmol.mL", "FDSLNH_nmol.mL", "DSLNH_nmol.mL",
                           "SUM_nmol.mL", "Sia_nmol.mL", "Fuc_nmol.mL"))) %>%
  drop_na()

XhmoE <- data.frame(hmoE %>%
                  select(c("nlchild_base", "nperson_base", "nrooms_base", 
                            "meducyrs_base", "h2osrcp_base", "cookplac_base",
                            "inctot_base")))

XhmoE $ cookplac_base <- as.factor(ifelse(XhmoE $ cookplac_base == "Yes", 1, 0))

# Make a function for h2o transformation to avoid small sample sizes
h2oFunc <- function(data, feature) {
  data[[feature]] <- as.factor(case_when(data[[feature]] == "Tube well or borehole" |
                              data[[feature]] == "Protected well" ~ 1,

                              data[[feature]] == "Surface water" |
                              data[[feature]] == "Unprotected well" ~ 2,

                              data[[feature]] == "Piped to yard/plot" |
                              data[[feature]] == "Public tap/stand pipe" ~ 3))
}

XhmoE $ h2osrcp_base = h2oFunc(data = XhmoE, feature = "h2osrcp_base")

YhmoE <- data.frame(hmoE %>%
                  select(c("X2.FL_nmol.mL", "X3FL_nmol.mL", "DFLac_nmol.mL",
                           "X3.SL_nmol.mL", "X6.SL_nmol.mL", "LNT_nmol.mL",
                           "LNnT_nmol.mL", "LNFP.I_nmol.mL", "LNFP.II_nmol.mL",
                           "LNFP.III_nmol.mL", "LSTb_nmol.mL", "LSTc_nmol.mL", 
                           "DFLNT_nmol.mL", "LNH_nmol.mL", "DSLNT_nmol.mL",
                           "FLNH_nmol.mL", "DFLNH_nmol.mL", "FDSLNH_nmol.mL",
                           "DSLNH_nmol.mL", "SUM_nmol.mL", "Sia_nmol.mL", "Fuc_nmol.mL")))

YhmoEscale = scale(YhmoE)

# Vitamin B Elicit
try(bvitE <- readRDS("/data/imic/data/raw_lab_data/elicit/merged_elicit/bvit.RDS"))
try(bvitE <- readRDS("C:/Users/andre/Downloads/bvit.RDS"))

# Prepare data for analysis
bvitE <- data.frame(bvitE %>%
                  select(c("nlchild_base", "nperson_base", "nrooms_base", 
                            "meducyrs_base", "h2osrcp_base", "cookplac_base",
                            "inctot_base", "B12", "TPP", "TMP", "T", "B1", "Ribo", "FMN", 
                            "FAD", "NAM", "NAD", "NMN", "NR", "Nufa", "PA", "PL", "PM", "PN", 
                            "PLP", "Bio", "TRP", "B2", "B3", "B6"))) %>%
  drop_na()

XbvitE <- data.frame(bvitE %>%
                  select(c("nlchild_base", "nperson_base", "nrooms_base", 
                            "meducyrs_base", "h2osrcp_base", "cookplac_base",
                            "inctot_base")))

XbvitE $ cookplac_base <- as.factor(ifelse(XbvitE $ cookplac_base == "Yes", 1, 0))

XbvitE $ h2osrcp_base = h2oFunc(data = XbvitE, feature = "h2osrcp_base")

YbvitE <- data.frame(bvitE %>%
                  select(c("B12", "TPP", "TMP", "T", "B1", "Ribo", "FMN", 
                            "FAD", "NAM", "NAD", "NMN", "NR", "Nufa", "PA", "PL",
                           "PM", "PN", "PLP", "Bio", "TRP", "B2", "B3", "B6")))

YbvitEscale = scale(YbvitE)

# Call the function on data
# Elicit HMO
hmoEresult <- r2weightFunc(X = XhmoE, Y = YhmoEscale)
predictors <- c("nlchild_base", "nperson_base", "nrooms_base", "meducyrs_base",
                "h2osrcp_base", "cookplac_base", "inctot_base")
hmoEresult $ predictors = predictors
# Save the dataset to directory
try(write.csv(hmoEresult, file = "/data/imic/results/r2weight/hmoEr2weight.csv"))
saveRDS(hmoEresult, file=paste0(here::here(),"/results/hmoEr2weight.RDS"))

# Elicit Bvit
bvitResult <- r2weightFunc(X = XbvitE, Y = YbvitEscale)
bvitResult $ predictors = predictors

try(write.csv(bvitResult, file = "/data/imic/results/r2weight/bvitEr2weight.csv"))
saveRDS(bvitResult, file=paste0(here::here(),"/results/bvitEr2weight.RDS"))

# We could also get importance for each individual outcome:
# call optWeight using simple Super Learner library
out1 <- optWeight(Y = YbvitEscale, X = XbvitE, SL.library = c("SL.mean","SL.glm"))#,"SL.step"))

# Measure importance of x7
out2 <- optWeight(Y = YbvitEscale, X = XbvitE[, 1:6], SL.library = c("SL.glm","SL.mean"))#,"SL.step"))

# Compare to full fit to get importance for each individual outcome
outDiff <- r2_diff(out1, out2)

# Difference in R-squared for first outcome with confidence interval and p-value
# for two-sided test that the difference equals 0
outDiff $ y1 $ diff

# For the third outcome
outDiff $ y3 $ diff
